requirements.txt
fastapi==0.115.0
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.6
supabase==2.11.0
pinecone-client==3.2.2
langchain==0.3.14
langchain-google-genai==2.0.7
langchain-huggingface==0.1.2
sentence-transformers==3.3.1
python-dotenv==1.0.1
pydantic==2.10.3
python-multipart==0.0.9
